requirements.txt
tensorflow==2.15.0
Flask==3.0.3
numpy==1.26.4
Pillow==10.3.0